* EasyBudget.Business.csproj: moved view model classes
* EasyBudgetDataService.cs:
* BaseViewModel.cs:
* DepositViewModel.cs:
* BudgetItemViewModel.cs:
* IncomeItemViewModel.cs:
* WithdrawalViewModel.cs:
* BankAccountViewModel.cs:
* BudgetItemsViewModel.cs:
* ExpenseItemViewModel.cs:
* BankAccountsViewModel.cs:
* BudgetCategoryViewModel.cs:
* SavingsAccountViewModel.cs:
* CheckingAccountViewModel.cs:
* BudgetCategoriesViewModel.cs:
* SavingsWithdrawalViewModel.cs:
* CheckingWithdrawalViewModel.cs:
