+++ EasyBudget.Business/EasyBudgetDataService.cs
@@ -19,6 +19,7 @@ using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Text;
 using System.Threading.Tasks;
+using EasyBudget.Business.ViewModels;
 using EasyBudget.Models;
 using EasyBudget.Models.DataModels;
 
@@ -59,71 +60,71 @@ namespace EasyBudget.Business
             }
         }
 
-        public async Task<BankAccountsVM> GetBankAccountsViewModelAsync()
+        public async Task<BankAccountsViewModel> GetBankAccountsViewModelAsync()
         {
-            BankAccountsVM vm = new BankAccountsVM(this.dbFilePath);
+            BankAccountsViewModel vm = new BankAccountsViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadBankAccountsAsync();
             return vm;
         }
 
-        public async Task<BudgetCategoriesVM> GetBudgetCategoriesViewModelAsync()
+        public async Task<BudgetCategoriesViewModel> GetBudgetCategoriesViewModelAsync()
         {
-            BudgetCategoriesVM vm = new BudgetCategoriesVM(this.dbFilePath);
+            BudgetCategoriesViewModel vm = new BudgetCategoriesViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadBudgetCategoriesAsync();
 
             return vm;
         }
 
-        public async Task<BudgetCategoryVM> GetBudgetCategoryVM(int categoryId)
+        public async Task<BudgetCategoryViewModel> GetBudgetCategoryVM(int categoryId)
         {
-            BudgetCategoryVM vm = new BudgetCategoryVM(this.dbFilePath);
+            BudgetCategoryViewModel vm = new BudgetCategoryViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadBudgetCategoryDetails(categoryId);
 
             return vm;
         }
 
-        public async Task<BudgetItemsVM> GetBudgetItemsVM(int categoryId)
+        public async Task<BudgetItemsViewModel> GetBudgetItemsVM(int categoryId)
         {
-            BudgetItemsVM vm = new BudgetItemsVM(this.dbFilePath);
+            BudgetItemsViewModel vm = new BudgetItemsViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadBudgetItemsAsync(categoryId);
 
             return vm;
         }
 
-        public async Task<CheckingAccountVM> GetCheckingAccountVMAsync(int accountId)
+        public async Task<CheckingAccountViewModel> GetCheckingAccountVMAsync(int accountId)
         {
-            CheckingAccountVM vm = new CheckingAccountVM(this.dbFilePath);
+            CheckingAccountViewModel vm = new CheckingAccountViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadCheckingAccountDetailsAsync(accountId);
 
             return vm;
         }
 
-        public async Task<SavingsAccountVM> GetSavingsAccountVMAsync(int accountId)
+        public async Task<SavingsAccountViewModel> GetSavingsAccountVMAsync(int accountId)
         {
-            SavingsAccountVM vm = new SavingsAccountVM(this.dbFilePath);
+            SavingsAccountViewModel vm = new SavingsAccountViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadSavingsAccountDetailsAsync(accountId);
 
             return vm;
         }
 
-        public async Task<IncomeItemVM> GetIncomeItemVMAsync(int itemId)
+        public async Task<IncomeItemViewModel> GetIncomeItemVMAsync(int itemId)
         {
-            IncomeItemVM vm = new IncomeItemVM(this.dbFilePath);
+            IncomeItemViewModel vm = new IncomeItemViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadIncomeItemAsync(itemId);
 
             return vm;
         }
 
-        public async Task<ExpenseItemVM> GetExpenseItemVMAsync(int itemId)
+        public async Task<ExpenseItemViewModel> GetExpenseItemVMAsync(int itemId)
         {
-            ExpenseItemVM vm = new ExpenseItemVM(this.dbFilePath);
+            ExpenseItemViewModel vm = new ExpenseItemViewModel(this.dbFilePath);
             await this.EnsureSystemItemsExistAsync();
             await vm.LoadExpenseItemAsync(itemId);
 
@@ -157,1583 +158,28 @@ namespace EasyBudget.Business
         //    return _success;
         //}
     
-        public BudgetCategoryVM CreateBudgetCategoryVM()
+        public BudgetCategoryViewModel CreateBudgetCategoryVM()
         {
-            var vm = new BudgetCategoryVM(this.dbFilePath);
+            var vm = new BudgetCategoryViewModel(this.dbFilePath);
             vm.CreateBudgetCategory();
 
             return vm;
         }
 
-        public CheckingAccountVM CreateCheckingAccountVM()
+        public CheckingAccountViewModel CreateCheckingAccountVM()
         {
-            var vm = new CheckingAccountVM(this.dbFilePath);
+            var vm = new CheckingAccountViewModel(this.dbFilePath);
             vm.CreateCheckingAccount();
 
             return vm;
         }
 
-        public SavingsAccountVM CreateSavingsAccountVM()
+        public SavingsAccountViewModel CreateSavingsAccountVM()
         {
-            var vm = new SavingsAccountVM(this.dbFilePath);
+            var vm = new SavingsAccountViewModel(this.dbFilePath);
             vm.CreateSavingsAccount();
 
             return vm;
-        }
-    }
-
-    public abstract class BaseViewModel
-    {
-        internal string dbFilePath;
-
-        public string ErrorCondition { get; set; }
-        
-        internal BaseViewModel(string dbFilePath)
-        {
-            this.dbFilePath = dbFilePath;
-        }
-
-        internal void WriteErrorCondition(string error)
-        {
-            if (string.IsNullOrEmpty(this.ErrorCondition))
-                this.ErrorCondition = error;
-            else
-            {
-                StringBuilder sb = new StringBuilder(this.ErrorCondition);
-                sb.AppendLine(error);
-                this.ErrorCondition = sb.ToString();
-            }
-        }
-
-        internal void WriteErrorCondition(Exception ex)
-        {
-            if (string.IsNullOrEmpty(this.ErrorCondition))
-                this.ErrorCondition = ex.Message;
-            else
-            {
-                StringBuilder sb = new StringBuilder(this.ErrorCondition);
-                sb.AppendLine(ex.Message);
-                this.ErrorCondition = sb.ToString();
-            }
-            if (ex.InnerException != null)
-            {
-                WriteErrorCondition(ex.InnerException);
-            }
-        }
-    }
-
-    public class BankAccountsVM : BaseViewModel
-    {
-
-        public ICollection<BankAccountVM> BankAccountVMs { get; set; }
-
-        internal BankAccountsVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-            BankAccountVMs = new List<BankAccountVM>();
-        }
-
-        internal async Task LoadBankAccountsAsync()
-        {
-            await LoadCheckingAccountsAsync();
-            await LoadSavingsAccountsAsync();
-        }
-
-        private async Task LoadCheckingAccountsAsync()
-        {
-
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetAllCheckingAccountsAsync();
-                if (_results.Successful)
-                {
-                    foreach (var account in _results.Results)
-                    {
-                        //this.BankAccountVMs.Add(account);
-                    }
-                }
-                else
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred");
-                    }
-                }
-            }
-        }
-
-        private async Task LoadSavingsAccountsAsync()
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetAllSavingsAccountsAsync();
-                if (_results.Successful)
-                {
-                    foreach (var account in _results.Results)
-                    {
-                        //this.BankAccountVMs.Add(account);
-                    }
-                }
-                else
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred");
-                    }
-                }
-            }
-        }
-
-    }
-
-    public class BankAccountVM : BaseViewModel
-    {
-        BankAccount Account { get; set; }
-
-        public string BankName { get; set; }
-
-        public BankAccountType AccountType { get; set; }
-
-        public decimal CurrentBalance { get; set; }
-
-        public string RoutingNumber { get; set; }
-
-        public string AccountNumber { get; set; }
-
-        public string Nickname { get; set; }
-
-        internal BankAccountVM(string dbFilePath)
-            :base(dbFilePath)
-        {
-            
-        }
-
-        internal async Task PopulateVMAsync(int id, BankAccountType accountType)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                switch (accountType)
-                {
-                    case BankAccountType.Checking:
-                        var _resultsChecking = await uow.GetCheckingAccountAsync(id);
-                        if (_resultsChecking.Successful)
-                        {
-                            this.Account = _resultsChecking.Results;
-                        }
-                        else
-                        {
-                            if (_resultsChecking.WorkException != null)
-                            {
-                                WriteErrorCondition(_resultsChecking.WorkException);
-                            }
-                            else if (!string.IsNullOrEmpty(_resultsChecking.Message))
-                            {
-                                WriteErrorCondition(_resultsChecking.Message);
-                            }
-                            else
-                            {
-                                WriteErrorCondition("An unknown error has occurred");
-                            }
-                        }
-                        break;
-                    case BankAccountType.Savings:
-                        var _resultsSavings = await uow.GetSavingsAccountAsync(id);
-                        if (_resultsSavings.Successful)
-                        {
-                            this.Account = _resultsSavings.Results;
-                        }
-                        else
-                        {
-                            if (_resultsSavings.WorkException != null)
-                            {
-                                WriteErrorCondition(_resultsSavings.WorkException);
-                            }
-                            else if (!string.IsNullOrEmpty(_resultsSavings.Message))
-                            {
-                                WriteErrorCondition(_resultsSavings.Message);
-                            }
-                            else
-                            {
-                                WriteErrorCondition("An unknown error has occurred");
-                            }
-                        }
-                        break;
-                }
-            }
-            if (this.Account != null)
-            {
-                this.BankName = this.Account.bankName;
-                this.AccountNumber = this.Account.accountNumber;
-                this.RoutingNumber = this.Account.routingNumber;
-                this.AccountType = this.Account.accountType;
-                this.CurrentBalance = this.Account.currentBalance;
-
-            }
-        }
-    }
-
-    public class DepositVM : BaseViewModel
-    {
-        public int accountId { get; set; }
-
-        public BankAccountType accountType { get; set; }
-
-        public DateTime transactionDate { get; set; }
-
-        public decimal transactionAmount { get; set; }
-
-        public string description { get; set; }
-
-        public string notation { get; set; }
-
-        public int budgetItemId { get; set; }
-
-        public bool reconciled { get; set; }
-
-        internal DepositVM(string dbFilePath)
-            :base(dbFilePath)
-        {
-            
-        }
-
-        internal void PopulateVM(CheckingDeposit deposit)
-        {
-            this.accountId = deposit.checkingAccountId;
-            this.transactionDate = deposit.transactionDate;
-            this.transactionAmount = deposit.transactionAmount;
-            this.description = deposit.description;
-            this.notation = deposit.notation;
-            this.budgetItemId = deposit.budgetIncomeId != null ? (int)deposit.budgetIncomeId : 0;
-            this.reconciled = deposit.reconciled;
-            this.accountType = BankAccountType.Checking;
-        }
-
-        internal void PopulateVM(SavingsDeposit deposit)
-        {
-            this.accountId = deposit.savingsAccountId;
-            this.transactionDate = deposit.transactionDate;
-            this.transactionAmount = deposit.transactionAmount;
-            this.description = deposit.description;
-            this.notation = deposit.notation;
-            this.budgetItemId = deposit.budgetIncomeId != null ? (int)deposit.budgetIncomeId : 0;
-            this.reconciled = deposit.reconciled;
-            this.accountType = BankAccountType.Savings;
-        }
-    }
-
-    public class WithdrawalVM : BaseViewModel
-    {
-        internal WithdrawalVM(string dbFilePath) 
-            : base(dbFilePath)
-        {
-            
-        }
-
-    }
-
-    public class BudgetItemsVM : BaseViewModel
-    {
-
-        List<BudgetItem> BudgetItems { get; set; }
-
-        ICollection<BudgetItemVM> BudgetItemVMs { get; set; }
-
-        internal BudgetItemsVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-            BudgetItems = new List<BudgetItem>();
-            BudgetItemVMs = new List<BudgetItemVM>();
-        }
-
-        internal async Task LoadBudgetItemsAsync(int categoryId)
-        {
-            await LoadIncomeItemsAsync(categoryId);
-            await LoadExpenseItemsAsync(categoryId);
-        }
-
-        private async Task LoadIncomeItemsAsync(int categoryId)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _resultsCategory = await uow.GetBudgetCategoryAsync(categoryId);
-                if (_resultsCategory.Successful)
-                {
-                    BudgetCategory category = _resultsCategory.Results;
-                    var _resultsIncomeItems = await uow.GetCategoryIncomeItemsAsync(category);
-                    if (_resultsIncomeItems.Successful)
-                    {
-                        foreach (var item in _resultsIncomeItems.Results)
-                        {
-                            item.ItemType = BudgetItemType.Income;
-                            this.BudgetItems.Add(item);
-                            var vm = new BudgetItemVM(this.dbFilePath);
-                            await vm.PopulateVMAsync(item.id, item.ItemType);
-                            this.BudgetItemVMs.Add(vm);
-                        }
-                    }
-                    else
-                    {
-                        if (_resultsIncomeItems.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsIncomeItems.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsIncomeItems.Message))
-                        {
-                            WriteErrorCondition(_resultsIncomeItems.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-                    }
-                }
-                else
-                {
-                    if (_resultsCategory.WorkException != null)
-                    {
-                        WriteErrorCondition(_resultsCategory.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_resultsCategory.Message))
-                    {
-                        WriteErrorCondition(_resultsCategory.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred");
-                    }
-                }
-            }
-        }
-
-        private async Task LoadExpenseItemsAsync(int categoryId)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _resultsCategory = await uow.GetBudgetCategoryAsync(categoryId);
-                if (_resultsCategory.Successful)
-                {
-                    BudgetCategory category = _resultsCategory.Results;
-                    var _resultsIncomeItems = await uow.GetCategoryExpenseItemsAsync(category);
-                    if (_resultsIncomeItems.Successful)
-                    {
-                        foreach (var item in _resultsIncomeItems.Results)
-                        {
-                            item.ItemType = BudgetItemType.Expense;
-                            this.BudgetItems.Add(item);
-                            var vm = new BudgetItemVM(this.dbFilePath);
-                            await vm.PopulateVMAsync(item.id, item.ItemType);
-                            this.BudgetItemVMs.Add(vm);
-                        }
-                    }
-                    else
-                    {
-                        if (_resultsIncomeItems.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsIncomeItems.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsIncomeItems.Message))
-                        {
-                            WriteErrorCondition(_resultsIncomeItems.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred loading Expense Items");
-                        }
-                    }
-                }
-                else
-                {
-                    if (_resultsCategory.WorkException != null)
-                    {
-                        WriteErrorCondition(_resultsCategory.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_resultsCategory.Message))
-                    {
-                        WriteErrorCondition(_resultsCategory.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred");
-                    }
-                }
-            }
-        }
-    
-    }
-
-    public class BudgetItemVM : BaseViewModel
-    {
-        BudgetItem source { get; set; }
-
-        public int CategoryId { get; set; }
-
-        public int BudgetItemId { get; set; }
-
-        public string CategoryName { get; set; }
-
-        public BudgetItemType ItemType { get; set; }
-
-        public decimal BudgetedAmount { get; set; }
-
-        public string ItemDescription { get; set; }
-
-        public string ItemNotation { get; set; }
-
-        public bool IsRecurring { get; set; }
-
-        public Frequency ItemFrequency { get; set; }
-
-        public DateTime StartDate { get; set; }
-
-        public DateTime EndDate { get; set; }
-
-        internal BudgetItemVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-
-        }
-
-        internal async Task PopulateVMAsync(int id, BudgetItemType itemType)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                switch (itemType)
-                {
-                    case BudgetItemType.Expense:
-                        var _resultsExpenseItem = await uow.GetExpenseItemAsync(id);
-                        if (_resultsExpenseItem.Successful)
-                        {
-                            this.source = _resultsExpenseItem.Results;
-                        }
-                        else
-                        {
-                            if (_resultsExpenseItem.WorkException != null)
-                            {
-                                WriteErrorCondition(_resultsExpenseItem.WorkException);
-                            }
-                            else if (!string.IsNullOrEmpty(_resultsExpenseItem.Message))
-                            {
-                                WriteErrorCondition(_resultsExpenseItem.Message);
-                            }
-                            else
-                            {
-                                WriteErrorCondition("An unknown error has occurred");
-                            }
-                        }
-                        break;
-                    case BudgetItemType.Income:
-                        var _resultsIncomeItem = await uow.GetIncomeItemAsync(id);
-                        if (_resultsIncomeItem.Successful)
-                        {
-                            this.source = _resultsIncomeItem.Results;
-                        }
-                        else
-                        {
-                            if (_resultsIncomeItem.WorkException != null)
-                            {
-                                WriteErrorCondition(_resultsIncomeItem.WorkException);
-                            }
-                            else if (!string.IsNullOrEmpty(_resultsIncomeItem.Message))
-                            {
-                                WriteErrorCondition(_resultsIncomeItem.Message);
-                            }
-                            else
-                            {
-                                WriteErrorCondition("An unknown error has occurred");
-                            }
-                        }
-                        break;
-                }
-
-                if (this.source != null)
-                {
-                    var _resultsCategory = await uow.GetBudgetCategoryAsync(this.source.budgetCategoryId);
-                    if (_resultsCategory.Successful)
-                    {
-                        this.source.budgetCategory = _resultsCategory.Results;
-                        this.CategoryName = this.source.budgetCategory.categoryName;
-                        this.BudgetedAmount = this.source.BudgetedAmount;
-                        this.BudgetItemId = this.source.id;
-                        this.EndDate = this.source.EndDate;
-                        this.StartDate = this.source.StartDate;
-                        this.IsRecurring = this.source.recurring;
-                        this.ItemDescription = this.source.description;
-                        this.ItemFrequency = this.source.frequency;
-                        this.ItemNotation = this.source.notation;
-                        this.ItemType = this.source.ItemType;
-
-                    }
-                    else
-                    {
-                        if (_resultsCategory.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsCategory.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsCategory.Message))
-                        {
-                            WriteErrorCondition(_resultsCategory.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    public class BudgetCategoriesVM : BaseViewModel
-    {
-
-        public ICollection<BudgetCategoryVM> BudgetCategoryVMs { get; set; }
-
-        public BudgetCategoriesVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-            this.BudgetCategoryVMs = new List<BudgetCategoryVM>();
-        }
-
-        internal async Task LoadBudgetCategoriesAsync()
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetAllBudgetCategoriesAsync();
-                if (_results.Successful)
-                {
-                    foreach (var category in _results.Results)
-                    {
-                        var vm = new BudgetCategoryVM(this.dbFilePath);
-                        await vm.PopulateVMAsync(category);
-                        this.BudgetCategoryVMs.Add(vm);
-                    }
-                }
-                else
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred");
-                    }
-                }
-            }
-        }
-
-    }
-
-    public class BudgetCategoryVM : BaseViewModel
-    {
-        BudgetCategory Category { get; set; }
-
-        public int CategoryId { get; set; }
-
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-        public decimal Amount { get; set; }
-
-        public bool IsSystemCategory { get; set; }
-
-        public bool IsUserSelected { get; set; }
-
-        public AppIcon CategoryIcon { get; set; }
-
-        public BudgetCategoryType CategoryType { get; set; }
-
-        public ICollection<BudgetItemVM> BudgetItemVMs { get; set; }
-
-        public bool IsNew { get; set; }
-
-        public BudgetCategoryVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-            this.BudgetItemVMs = new List<BudgetItemVM>();
-        }
-
-        internal async  Task PopulateVMAsync(BudgetCategory category)
-        {
-            
-            this.Category = category;
-            this.CategoryId = category.id;
-            this.Name = category.categoryName;
-            this.Description = category.description;
-            this.Amount = category.budgetAmount;
-            this.IsSystemCategory = category.systemCategory;
-            this.IsUserSelected = category.userSelected;
-            this.CategoryIcon = category.categoryIcon;
-            this.CategoryType = category.categoryType;
-            this.IsNew = category.IsNew;
-
-            using (UnitOfWork uow = new UnitOfWork(dbFilePath))
-            {
-                switch (category.categoryType)
-                {
-                    case BudgetCategoryType.Expense:
-                        var _resultsExpenseItems = await uow.GetCategoryExpenseItemsAsync(category);
-                        if (_resultsExpenseItems.Successful)
-                        {
-                            foreach (var item in _resultsExpenseItems.Results)
-                            {
-                                item.ItemType = BudgetItemType.Expense;
-                                item.budgetCategory = category;
-                                var vm = new BudgetItemVM(this.dbFilePath);
-                                await vm.PopulateVMAsync(item.id, item.ItemType);
-                                this.BudgetItemVMs.Add(vm);
-                            }
-                        }
-                        else
-                        {
-                            if (_resultsExpenseItems.WorkException != null)
-                            {
-                                WriteErrorCondition(_resultsExpenseItems.WorkException);
-                            }
-                            else if (!string.IsNullOrEmpty(_resultsExpenseItems.Message))
-                            {
-                                WriteErrorCondition(_resultsExpenseItems.Message);
-                            }
-                            else
-                            {
-                                WriteErrorCondition("An unknown error has occurred loading Expense Items");
-                            }
-                        }
-                        break;
-                    case BudgetCategoryType.Income:
-                        var _resultsIncomeItems = await uow.GetCategoryIncomeItemsAsync(category);
-                        if (_resultsIncomeItems.Successful)
-                        {
-                            foreach (var item in _resultsIncomeItems.Results)
-                            {
-                                item.ItemType = BudgetItemType.Income;
-                                item.budgetCategory = category;
-                                var vm = new BudgetItemVM(this.dbFilePath);
-                                await vm.PopulateVMAsync(item.id, item.ItemType);
-                                this.BudgetItemVMs.Add(vm);
-                            }
-                        }
-                        else
-                        {
-                            if (_resultsIncomeItems.WorkException != null)
-                            {
-                                WriteErrorCondition(_resultsIncomeItems.WorkException);
-                            }
-                            else if (!string.IsNullOrEmpty(_resultsIncomeItems.Message))
-                            {
-                                WriteErrorCondition(_resultsIncomeItems.Message);
-                            }
-                            else
-                            {
-                                WriteErrorCondition("An unknown error has occurred loading Income Items");
-                            }
-                        }
-                        break;
-                }
-            }
-
-        }
-
-
-        internal async Task LoadBudgetCategoryDetails(int categoryId)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetBudgetCategoryAsync(categoryId);
-                if (_results.Successful)
-                {
-                    await PopulateVMAsync(_results.Results);
-                }
-                else
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred loading the Budget Category");
-                    }
-                }
-            }
-        }
-
-        internal void CreateBudgetCategory()
-        {
-            this.Category = new BudgetCategory();
-            this.Category.IsNew = true;
-
-
-        }
-
-        public async Task SaveChangesAsync()
-        {
-            bool _saveOk = true;
-            this.Category.categoryName = this.Name;
-            this.Category.description = this.Description;
-            this.Category.categoryType = this.CategoryType;
-            this.Category.budgetAmount = this.Amount;
-            this.Category.systemCategory = this.IsSystemCategory;
-            this.Category.categoryIcon = this.CategoryIcon;
-            this.Category.userSelected = this.IsUserSelected;
-
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                if (this.IsNew)
-                {
-                    
-                    var _resultsAddCategory = await uow.AddBudgetCategoryAsync(this.Category);
-                    if (_resultsAddCategory.Successful)
-                    {
-                        _saveOk = true;
-                        this.CategoryId = _resultsAddCategory.Results.id;
-                        this.IsNew = false;
-                    }
-                    else
-                    {
-                        _saveOk = false;
-                        if (_resultsAddCategory.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsAddCategory.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsAddCategory.Message))
-                        {
-                            WriteErrorCondition(_resultsAddCategory.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-                    }
-                }
-                else
-                {
-                    var _resultsUpdateCategory = await uow.UpdateBudgetCategoryAsync(this.Category);
-                    if (_resultsUpdateCategory.Successful)
-                    {
-                        _saveOk = true;
-                        this.Category.id = _resultsUpdateCategory.Results.id;
-                        this.IsNew = false;
-                    }
-                    else
-                    {
-                        _saveOk = false;
-                        if (_resultsUpdateCategory.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsUpdateCategory.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsUpdateCategory.Message))
-                        {
-                            WriteErrorCondition(_resultsUpdateCategory.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-
-                    }
-                }
-                if (_saveOk)
-                {
-                    
-                    //foreach (IncomeItem itm in this.IncomeItems)
-                    //{
-                        //if (itm.IsNew)
-                        //{
-                        //    var _resultsAddItem = await uow.AddIncomeItemAsync(itm);
-                        //    if (_resultsAddItem.Successful)
-                        //    {
-                        //        _saveOk = _saveOk && true;
-                        //    }
-                        //    else
-                        //    {
-                        //        _saveOk = false;
-                        //        if (_resultsAddItem.WorkException != null)
-                        //        {
-                        //            WriteErrorCondition(_resultsAddItem.WorkException);
-                        //        }
-                        //        else if (!string.IsNullOrEmpty(_resultsAddItem.Message))
-                        //        {
-                        //            WriteErrorCondition(_resultsAddItem.Message);
-                        //        }
-                        //        else
-                        //        {
-                        //            WriteErrorCondition("An unknown error has occurred");
-                        //        }
-                        //    }
-                        //}
-                        //else
-                        //{
-                        //    var _resultsUpdateItem = await uow.UpdateIncomeItemAsync(itm);
-                        //    if (_resultsUpdateItem.Successful)
-                        //    {
-                        //        _saveOk = _saveOk && true;
-                        //    }
-                        //    else
-                        //    {
-                        //        _saveOk = false;
-                        //        if (_resultsUpdateItem.WorkException != null)
-                        //        {
-                        //            WriteErrorCondition(_resultsUpdateItem.WorkException);
-                        //        }
-                        //        else if (!string.IsNullOrEmpty(_resultsUpdateItem.Message))
-                        //        {
-                        //            WriteErrorCondition(_resultsUpdateItem.Message);
-                        //        }
-                        //        else
-                        //        {
-                        //            WriteErrorCondition("An unknown error has occurred");
-                        //        }
-                        //    }
-                        //}
-                    //}
-                    //foreach (ExpenseItem itm in this.ExpenseItems)
-                    //{
-                    //    if (itm.IsNew)
-                    //    {
-                    //        var _resultsAddItem = await uow.AddExpenseItemAsync(itm);
-                    //        if (_resultsAddItem.Successful)
-                    //        {
-                    //            _saveOk = _saveOk && true;
-                    //        }
-                    //        else
-                    //        {
-                    //            _saveOk = false;
-                    //            if (_resultsAddItem.WorkException != null)
-                    //            {
-                    //                WriteErrorCondition(_resultsAddItem.WorkException);
-                    //            }
-                    //            else if (!string.IsNullOrEmpty(_resultsAddItem.Message))
-                    //            {
-                    //                WriteErrorCondition(_resultsAddItem.Message);
-                    //            }
-                    //            else
-                    //            {
-                    //                WriteErrorCondition("An unknown error has occurred");
-                    //            }
-                    //        }
-                    //    }
-                    //    else
-                    //    {
-                    //        var _resultsUpdateItem = await uow.UpdateExpenseItemAsync(itm);
-                    //        if (_resultsUpdateItem.Successful)
-                    //        {
-                    //            _saveOk = _saveOk && true;
-                    //        }
-                    //        else
-                    //        {
-                    //            _saveOk = false;
-                    //            if (_resultsUpdateItem.WorkException != null)
-                    //            {
-                    //                WriteErrorCondition(_resultsUpdateItem.WorkException);
-                    //            }
-                    //            else if (!string.IsNullOrEmpty(_resultsUpdateItem.Message))
-                    //            {
-                    //                WriteErrorCondition(_resultsUpdateItem.Message);
-                    //            }
-                    //            else
-                    //            {
-                    //                WriteErrorCondition("An unknown error has occurred");
-                    //            }
-                    //        }
-                    //    }
-                    //}
-                }
-            }
-        }
-
-        public ExpenseItem AddExpenseItem()
-        {
-            ExpenseItem item = new ExpenseItem();
-            item.budgetCategoryId = this.Category.id;
-            item.IsNew = true;
-
-            //this.ExpenseItems.Add(item);
-
-            return item;
-        }
-
-        public IncomeItem AddIncomeItem()
-        {
-            IncomeItem item = new IncomeItem();
-            item.budgetCategoryId = this.Category.id;
-            item.IsNew = true;
-            //this.IncomeItems.Add(item);
-
-            return item;
-        }
-    }
-
-    public class CheckingAccountVM : BaseViewModel
-    {
-        CheckingAccount CheckingAccount { get; set; }
-
-        public string RoutingNumber { get; set; }
-
-        public string AccountNumber { get; set; }
-
-        public string BankName { get; set; }
-
-        public string AccountNickname { get; set; }
-
-        public decimal CurrentBalance { get; set; }
-
-        public DateTime LoadTransactionsFromDate { get; set; }
-
-        public DateTime LoadTransactionsToDate { get; set; }
-
-        public ICollection<CheckingWithdrawal> Withdrawals { get; set; }
-
-        public ICollection<CheckingDeposit> Deposits { get; set; }
-
-        public CheckingAccountVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-            this.Withdrawals = new List<CheckingWithdrawal>();
-            this.Deposits = new List<CheckingDeposit>();
-        }
-
-        internal async Task LoadCheckingAccountDetailsAsync(int accountId)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetCheckingAccountAsync(accountId);
-                if (_results.Successful)
-                {
-                    this.CheckingAccount = _results.Results;
-                    this.RoutingNumber = this.CheckingAccount?.routingNumber ?? string.Empty;
-                    this.AccountNumber = this.CheckingAccount?.accountNumber ?? string.Empty;
-                    this.BankName = this.CheckingAccount?.bankName ?? string.Empty;
-                    this.AccountNickname = this.CheckingAccount?.accountNickname ?? string.Empty;
-                    this.CurrentBalance = this.CheckingAccount?.currentBalance ?? 0;
-                    if (this.CheckingAccount.withdrawals != null)
-                    {
-                        foreach(CheckingWithdrawal item in this.CheckingAccount.withdrawals)
-                        {
-                            this.Withdrawals.Add(item);
-                        }
-                    }
-                    if (this.CheckingAccount.deposits != null)
-                    {
-                        foreach(CheckingDeposit item in this.CheckingAccount.deposits)
-                        {
-                            this.Deposits.Add(item);
-                        }
-                    }
-                }
-                else
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred loading Checking Account");
-                    }
-                }
-            }
-        }
-    
-        public async Task SaveChangesAsync()
-        {
-            bool _saveOk = true;
-
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                if (this.CheckingAccount.IsNew)
-                {
-                    var _resultsAddAccont = await uow.AddCheckingAccountAsync(this.CheckingAccount);
-                    if (_resultsAddAccont.Successful)
-                    {
-                        _saveOk = true;
-                    }
-                    else
-                    {
-                        _saveOk = false;
-                        if (_resultsAddAccont.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsAddAccont.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsAddAccont.Message))
-                        {
-                            WriteErrorCondition(_resultsAddAccont.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-                    }
-                }
-                else
-                {
-                    var _resultsUpdateAccont = await uow.UpdateCheckingAccountAsync(this.CheckingAccount);
-                    if (_resultsUpdateAccont.Successful)
-                    {
-                        _saveOk = true;
-                    }
-                    else
-                    {
-                        _saveOk = false;
-                        if (_resultsUpdateAccont.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsUpdateAccont.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsUpdateAccont.Message))
-                        {
-                            WriteErrorCondition(_resultsUpdateAccont.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-                    }
-                }
-                if (_saveOk)
-                {
-                    foreach(CheckingDeposit deposit in this.Deposits)
-                    {
-                        if (deposit.IsNew)
-                        {
-                            var _resultsAddDeposit = await uow.AddCheckingDepositAsync(deposit);
-                            if (_resultsAddDeposit.Successful)
-                            {
-                                _saveOk = _saveOk && true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsAddDeposit.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsAddDeposit.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsAddDeposit.Message))
-                                {
-                                    WriteErrorCondition(_resultsAddDeposit.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                        else
-                        {
-                            var _resultsUpdateDeposit = await uow.UpdateCheckingDepositAsync(deposit);
-                            if (_resultsUpdateDeposit.Successful)
-                            {
-                                _saveOk = true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsUpdateDeposit.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsUpdateDeposit.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsUpdateDeposit.Message))
-                                {
-                                    WriteErrorCondition(_resultsUpdateDeposit.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                    }
-                    foreach (CheckingWithdrawal withdrawal in this.Withdrawals)
-                    {
-                        if (withdrawal.IsNew)
-                        {
-                            var _resultsAddWithdrawal = await uow.AddCheckingWithdrawalAsync(withdrawal);
-                            if (_resultsAddWithdrawal.Successful)
-                            {
-                                _saveOk = true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsAddWithdrawal.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsAddWithdrawal.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsAddWithdrawal.Message))
-                                {
-                                    WriteErrorCondition(_resultsAddWithdrawal.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                        else 
-                        {
-                            var _resultsUpdateWithdrawal = await uow.UpdateCheckingWithdrawalAsync(withdrawal);
-                            if (_resultsUpdateWithdrawal.Successful)
-                            {
-                                _saveOk = true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsUpdateWithdrawal.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsUpdateWithdrawal.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsUpdateWithdrawal.Message))
-                                {
-                                    WriteErrorCondition(_resultsUpdateWithdrawal.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        internal void CreateCheckingAccount()
-        {
-            this.CheckingAccount = new CheckingAccount();
-            this.CheckingAccount.IsNew = true;
-        }
-
-        public CheckingDeposit AddDeposit()
-        {
-            CheckingDeposit deposit = new CheckingDeposit();
-            deposit.checkingAccountId = this.CheckingAccount.id;
-            deposit.IsNew = true;
-
-            this.Deposits.Add(deposit);
-
-            return deposit;
-        }
-
-        public CheckingWithdrawal AddWithdrawal()
-        {
-            CheckingWithdrawal withdrawal = new CheckingWithdrawal();
-            withdrawal.checkingAccountId = this.CheckingAccount.id;
-            withdrawal.IsNew = true;
-
-            this.Withdrawals.Add(withdrawal);
-
-            return withdrawal;
-        }
-    }
-
-    public class SavingsAccountVM : BaseViewModel
-    {
-        SavingsAccount SavingsAccount { get; set; }
-
-        public string RoutingNumber { get; set; }
-
-        public string AccountNumber { get; set; }
-
-        public string BankName { get; set; }
-
-        public string AccountNickname { get; set; }
-
-        public decimal CurrentBalance { get; set; }
-
-        public ICollection<SavingsWithdrawal> Withdrawals { get; set; }
-
-        public ICollection<SavingsDeposit> Deposits { get; set; }
-
-        public SavingsAccountVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-            this.Withdrawals = new List<SavingsWithdrawal>();
-            this.Deposits = new List<SavingsDeposit>();
-        }
-
-        internal async Task LoadSavingsAccountDetailsAsync(int accountId)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetSavingsAccountAsync(accountId);
-                if (_results.Successful)
-                {
-                    this.SavingsAccount = _results.Results;
-                    this.RoutingNumber = this.SavingsAccount?.routingNumber ?? string.Empty;
-                    this.AccountNumber = this.SavingsAccount?.accountNumber ?? string.Empty;
-                    this.BankName = this.SavingsAccount?.bankName ?? string.Empty;
-                    this.AccountNickname = this.SavingsAccount?.accountNickname ?? string.Empty;
-                    this.CurrentBalance = this.SavingsAccount?.currentBalance ?? 0;
-                    if (this.SavingsAccount.withdrawals != null)
-                    {
-                        foreach (SavingsWithdrawal item in this.SavingsAccount.withdrawals)
-                        {
-                            this.Withdrawals.Add(item);
-                        }
-                    }
-                    if (this.SavingsAccount.deposits != null)
-                    {
-                        foreach (SavingsDeposit item in this.SavingsAccount.deposits)
-                        {
-                            this.Deposits.Add(item);
-                        }
-                    }
-                }
-                else
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred loading Savings Account");
-                    }
-                }
-            }
-        }
-    
-        internal void CreateSavingsAccount()
-        {
-            this.SavingsAccount = new SavingsAccount();
-            this.SavingsAccount.IsNew = true;
-        }
-
-        public async Task SaveChangesAsync()
-        {
-            bool _saveOk = true;
-
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                if (this.SavingsAccount.IsNew)
-                {
-                    var _resultsAddAccont = await uow.AddSavingsAccountAsync(this.SavingsAccount);
-                    if (_resultsAddAccont.Successful)
-                    {
-                        _saveOk = true;
-                    }
-                    else
-                    {
-                        _saveOk = false;
-                        if (_resultsAddAccont.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsAddAccont.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsAddAccont.Message))
-                        {
-                            WriteErrorCondition(_resultsAddAccont.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-                    }
-                }
-                else
-                {
-                    var _resultsUpdateAccont = await uow.UpdateSavingsAccountAsync(this.SavingsAccount);
-                    if (_resultsUpdateAccont.Successful)
-                    {
-                        _saveOk = true;
-                    }
-                    else
-                    {
-                        _saveOk = false;
-                        if (_resultsUpdateAccont.WorkException != null)
-                        {
-                            WriteErrorCondition(_resultsUpdateAccont.WorkException);
-                        }
-                        else if (!string.IsNullOrEmpty(_resultsUpdateAccont.Message))
-                        {
-                            WriteErrorCondition(_resultsUpdateAccont.Message);
-                        }
-                        else
-                        {
-                            WriteErrorCondition("An unknown error has occurred");
-                        }
-                    }
-                }
-                if (_saveOk)
-                {
-                    foreach (SavingsDeposit deposit in this.Deposits)
-                    {
-                        if (deposit.IsNew)
-                        {
-                            var _resultsAddDeposit = await uow.AddSavingsDepositAsync(deposit);
-                            if (_resultsAddDeposit.Successful)
-                            {
-                                _saveOk = _saveOk && true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsAddDeposit.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsAddDeposit.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsAddDeposit.Message))
-                                {
-                                    WriteErrorCondition(_resultsAddDeposit.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                        else
-                        {
-                            var _resultsUpdateDeposit = await uow.UpdateSavingsDepositAsync(deposit);
-                            if (_resultsUpdateDeposit.Successful)
-                            {
-                                _saveOk = true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsUpdateDeposit.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsUpdateDeposit.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsUpdateDeposit.Message))
-                                {
-                                    WriteErrorCondition(_resultsUpdateDeposit.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                    }
-                    foreach (SavingsWithdrawal withdrawal in this.Withdrawals)
-                    {
-                        if (withdrawal.IsNew)
-                        {
-                            var _resultsAddWithdrawal = await uow.AddSavingsWithdrawalAsync(withdrawal);
-                            if (_resultsAddWithdrawal.Successful)
-                            {
-                                _saveOk = true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsAddWithdrawal.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsAddWithdrawal.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsAddWithdrawal.Message))
-                                {
-                                    WriteErrorCondition(_resultsAddWithdrawal.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                        else
-                        {
-                            var _resultsUpdateWithdrawal = await uow.UpdateSavingsWithdrawalAsync(withdrawal);
-                            if (_resultsUpdateWithdrawal.Successful)
-                            {
-                                _saveOk = true;
-                            }
-                            else
-                            {
-                                _saveOk = false;
-                                if (_resultsUpdateWithdrawal.WorkException != null)
-                                {
-                                    WriteErrorCondition(_resultsUpdateWithdrawal.WorkException);
-                                }
-                                else if (!string.IsNullOrEmpty(_resultsUpdateWithdrawal.Message))
-                                {
-                                    WriteErrorCondition(_resultsUpdateWithdrawal.Message);
-                                }
-                                else
-                                {
-                                    WriteErrorCondition("An unknown error has occurred");
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        public SavingsDeposit AddDeposit()
-        {
-            var deposit = new SavingsDeposit();
-            deposit.savingsAccountId = this.SavingsAccount.id;
-            deposit.IsNew = true;
-
-            this.Deposits.Add(deposit);
-
-            return deposit;
-        }
-
-        public SavingsWithdrawal AddWithdrawal()
-        {
-            var withdrawal = new SavingsWithdrawal();
-            withdrawal.savingsAccountId = this.SavingsAccount.id;
-            withdrawal.IsNew = true;
-
-            this.Withdrawals.Add(withdrawal);
-
-            return withdrawal;
-        }
-    }
-
-    public class ExpenseItemVM : BaseViewModel
-    {
-        
-        ExpenseItem ExpenseItem;
-
-        public decimal BudgetedAmount { get; set; }
-
-        public string Description { get; set; }
-
-        public string Notation { get; set; }
-
-        public bool Recurring { get; set; }
-
-        public ExpenseItemVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-            
-        }
-
-        internal async Task LoadExpenseItemAsync(int itemId)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetExpenseItemAsync(itemId);
-                if (_results.Successful)
-                {
-                    
-                    this.ExpenseItem = _results.Results;
-                    this.ExpenseItem.ItemType = BudgetItemType.Expense;
-                    this.BudgetedAmount = this.ExpenseItem?.BudgetedAmount ?? 0;
-                    this.Description = this.ExpenseItem?.description ?? string.Empty;
-                    this.Notation = this.ExpenseItem?.notation ?? string.Empty;
-                    this.Recurring = this.ExpenseItem?.recurring ?? false;
-
-                }
-                else 
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred loading Expense Item");
-                    }
-                }
-            }
-        }
-
-        internal void CrateExpenseItem(int categoryId)
-        {
-            this.ExpenseItem = new ExpenseItem();
-            this.ExpenseItem.budgetCategoryId = categoryId;
-            this.ExpenseItem.IsNew = true;
-        }
-    }
-
-    public class IncomeItemVM : BaseViewModel
-    {
-        IncomeItem IncomeItem;
-
-        public decimal BudgetedAmount { get; set; }
-
-        public string Description { get; set; }
-
-        public string Notation { get; set; }
-
-        public bool Recurring { get; set; }
-
-        public IncomeItemVM(string dbFilePath)
-            : base(dbFilePath)
-        {
-
-        }
-
-        internal async Task LoadIncomeItemAsync(int itemId)
-        {
-            using (UnitOfWork uow = new UnitOfWork(this.dbFilePath))
-            {
-                var _results = await uow.GetIncomeItemAsync(itemId);
-                if (_results.Successful)
-                {
-                    this.IncomeItem = _results.Results;
-                    this.IncomeItem.ItemType = BudgetItemType.Income;
-                    this.BudgetedAmount = this.IncomeItem?.BudgetedAmount ?? 0;
-                    this.Description = this.IncomeItem?.description ?? string.Empty;
-                    this.Notation = this.IncomeItem?.notation ?? string.Empty;
-                    this.Recurring = this.IncomeItem?.recurring ?? false;
-
-                }
-                else
-                {
-                    if (_results.WorkException != null)
-                    {
-                        WriteErrorCondition(_results.WorkException);
-                    }
-                    else if (!string.IsNullOrEmpty(_results.Message))
-                    {
-                        WriteErrorCondition(_results.Message);
-                    }
-                    else
-                    {
-                        WriteErrorCondition("An unknown error has occurred loading Income Item");
-                    }
-                }
-            }
-        }
-    
-        internal void CreateIncomeItem(int categoryId)
-        {
-            this.IncomeItem = new IncomeItem();
-            this.IncomeItem.budgetCategoryId = categoryId;
-            this.IncomeItem.IsNew = true;
         }
     }
