+++ EasyBudget.Business/ViewModels/BaseViewModel.cs
@@ -1,0 +1,61 @@
+﻿//
+//  Copyright 2018  CrawfordNET Solutions, LLC
+//
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+//
+//        http://www.apache.org/licenses/LICENSE-2.0
+//
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+using System;
+using System.Text;
+
+namespace EasyBudget.Business.ViewModels
+{
+
+    public abstract class BaseViewModel
+    {
+        internal string dbFilePath;
+
+        public string ErrorCondition { get; set; }
+
+        internal BaseViewModel(string dbFilePath)
+        {
+            this.dbFilePath = dbFilePath;
+        }
+
+        internal void WriteErrorCondition(string error)
+        {
+            if (string.IsNullOrEmpty(this.ErrorCondition))
+                this.ErrorCondition = error;
+            else
+            {
+                StringBuilder sb = new StringBuilder(this.ErrorCondition);
+                sb.AppendLine(error);
+                this.ErrorCondition = sb.ToString();
+            }
+        }
+
+        internal void WriteErrorCondition(Exception ex)
+        {
+            if (string.IsNullOrEmpty(this.ErrorCondition))
+                this.ErrorCondition = ex.Message;
+            else
+            {
+                StringBuilder sb = new StringBuilder(this.ErrorCondition);
+                sb.AppendLine(ex.Message);
+                this.ErrorCondition = sb.ToString();
+            }
+            if (ex.InnerException != null)
+            {
+                WriteErrorCondition(ex.InnerException);
+            }
+        }
+    }
+
+}
